Prometheus doesn't like it when you set custom timestamps
Originally, I tried to use the timestamp field to track the date given
by snapcraft for the metrics. Prometheus complained that the metrics
were too far out of the appropriate window, so instead add the date
given as a label and let prometheus set it's own metric about when it
was scraped.